MAINT Split conftest.py into modules (#2418)
