+++ docs/project/changelog.md
@@ -54,6 +54,10 @@ substitutions:
   translated to negative Python ints.
   {pr}`2484`
 
+- {{ BREAKING }} `pyodide_build.testing` is removed. `run_in_pyodide` decorator can now be accessed
+  through `pyodide_test_runner`.
+  {pr}`2418`
+
 - {{ Enhancement }} Added the `js_id` attribute to `JsProxy` to allow using
   JavaScript object identity as a dictionary key.
   {pr}`2515`
