+++ packages/cryptography/test_cryptography.py
@@ -1,8 +1,7 @@
 from hypothesis import HealthCheck, given, settings
 from hypothesis.strategies import binary, integers
-
-from conftest import selenium_context_manager
-from pyodide_build.testing import run_in_pyodide
+from pyodide_test_runner import run_in_pyodide
+from pyodide_test_runner.fixture import selenium_context_manager
 
 
 @run_in_pyodide(packages=["cryptography"])
