+++ packages/lazy-object-proxy/test_lazy_object_proxy.py
@@ -1,4 +1,4 @@
-from pyodide_build.testing import run_in_pyodide
+from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["lazy-object-proxy"])
