+++ packages/micropip/test/snowballstemmer-2.0.0-py2.py3-none-any.whl
