+++ packages/opencv-python/test_opencv_python.py
@@ -1,7 +1,7 @@
 import base64
 import pathlib
 
-from pyodide_build.testing import run_in_pyodide
+from pyodide_test_runner import run_in_pyodide
 
 REFERENCE_IMAGES_PATH = pathlib.Path(__file__).parent / "reference-images"
