+++ packages/scikit-learn/test_scikit-learn.py
@@ -1,6 +1,5 @@
 import pytest
-
-from conftest import selenium_context_manager
+from pyodide_test_runner.fixture import selenium_context_manager
 
 
 @pytest.mark.driver_timeout(40)
