+++ packages/scipy/test_scipy.py
@@ -1,6 +1,6 @@
-from pyodide_build import testing
+from pyodide_test_runner import run_in_pyodide
 
-run_in_pyodide = testing.run_in_pyodide(
+run_in_pyodide_scipy = run_in_pyodide(
     module_scope=True,
     packages=["scipy"],
     # xfail_browsers={"chrome": "Times out in chrome"},
@@ -8,7 +8,7 @@ run_in_pyodide = testing.run_in_pyodide(
 )
 
 
-@run_in_pyodide
+@run_in_pyodide_scipy
 def test_scipy_linalg():
     import numpy as np
     import scipy.linalg
@@ -24,28 +24,28 @@ def test_scipy_linalg():
     assert_allclose(res, np.identity(N), rtol=1e-07, atol=1e-9)
 
 
-@run_in_pyodide
+@run_in_pyodide_scipy
 def test_brentq():
     from scipy.optimize import brentq
 
     brentq(lambda x: x, -1, 1)
 
 
-@run_in_pyodide
+@run_in_pyodide_scipy
 def test_dlamch():
     from scipy.linalg import lapack
 
     lapack.dlamch("Epsilon-Machine")
 
 
-@run_in_pyodide
+@run_in_pyodide_scipy
 def test_binom_ppf():
     from scipy.stats import binom
 
     assert binom.ppf(0.9, 1000, 0.1) == 112
 
 
-@testing.run_in_pyodide(module_scope=True, packages=["pytest", "scipy-tests"])
+@run_in_pyodide(module_scope=True, packages=["pytest", "scipy-tests"])
 def test_scipy_pytest():
     import pytest
