+++ packages/wrapt/test_wrapt.py
@@ -1,4 +1,4 @@
-from pyodide_build.testing import run_in_pyodide
+from pyodide_test_runner import run_in_pyodide
 
 
 @run_in_pyodide(packages=["wrapt"])
