+++ pyodide-test-runner/pyodide_test_runner/node_test_driver.js
@@ -5,11 +5,14 @@ const util = require("util");
 const node_fetch = require("node-fetch");
 const base64 = require("base-64");
 
-let { loadPyodide } = require("pyodide");
-let base_url = process.argv[2];
+let baseUrl = process.argv[2];
+let distDir = process.argv[3];
+
+let { loadPyodide } = require(`${distDir}/pyodide`);
+
 // node requires full paths.
 function fetch(path) {
-  return node_fetch(new URL(path, base_url).toString());
+  return node_fetch(new URL(path, baseUrl).toString());
 }
 
 const context = {
