+++ pyodide-test-runner/pyodide_test_runner/tests/test_testing.py
@@ -1,7 +1,8 @@
 import pathlib
 from textwrap import dedent
 
-from pyodide_build.testing import _run_in_pyodide_get_source, run_in_pyodide
+from pyodide_test_runner import run_in_pyodide
+from pyodide_test_runner.decorator import _run_in_pyodide_get_source
 
 
 def test_web_server_secondary(selenium, web_server_secondary):
@@ -89,7 +90,6 @@ def test_assert(selenium):
         let threw;
         assertThrows(() => { throw new TypeError("aaabbbccc") }, "TypeError", "bbc");
         assertThrows(() => { throw new TypeError("aaabbbccc") }, "TypeError", /.{3}.{3}.{3}/);
-
         threw = false;
         try {
             assertThrows(() => 0, "TypeError", /.*/);
@@ -98,7 +98,6 @@ def test_assert(selenium):
             assert(() => e.message == `assertThrows(() => 0, "TypeError", /.*/) failed, no error thrown`, e.message);
         }
         assert(() => threw);
-
         threw = false;
         try {
             assertThrows(() => { throw new ReferenceError("blah"); }, "TypeError", /.*/);
@@ -107,7 +106,6 @@ def test_assert(selenium):
             assert(() => e.message.endsWith("expected error of type 'TypeError' got type 'ReferenceError'"));
         }
         assert(() => threw);
-
         threw = false;
         try {
             assertThrows(() => { throw new TypeError("blah"); }, "TypeError", "abcd");
@@ -117,7 +115,6 @@ def test_assert(selenium):
             assert(() => e.message.endsWith(`expected error message to match pattern "abcd" got:\nblah`));
         }
         assert(() => threw);
-
         threw = false;
         try {
             assertThrows(() => { throw new TypeError("blah"); }, "TypeError", /a..d/);
