+++ src/tests/test_pyodide.py
@@ -3,9 +3,9 @@ from textwrap import dedent
 from typing import Any, Sequence
 
 import pytest
+from pyodide_test_runner import run_in_pyodide
 
 from pyodide import CodeRunner, eval_code, find_imports, should_quiet  # noqa: E402
-from pyodide_build.testing import run_in_pyodide
 
 
 def _strip_assertions_stderr(messages: Sequence[str]) -> list[str]:
